LayoutTests/editing/smart-lists/different-list-styles-do-not-merge.html

<!DOCTYPE html><!-- webkit-test-runner [ SmartListsAvailable=true ] -->
<html>
<head>
<script src="../editing.js"></script>
<script src="../../resources/dump-as-markup.js"></script>
</head>
<body>
<div id="test" contenteditable></div>
<script>
Markup.description('Adjacent lists with different marker styles (disc and dash) should not merge into a single list. After a blank line, typing a different marker type produces a separate list.');

var test = document.getElementById('test');
test.focus();
[...'* A\n* B\nC\n\n* D\n* E\n\n- A\n- B'].forEach(typeCharacterCommand);
Markup.dump('test');
</script>
</body>
</html>
